fix(ariadne): §9 volren/volslice review — dup-id, TF window, blank-guard, double domain
Adversarial review of the volren/volslice diff confirmed 5 findings (all real):

- Duplicate top-level id not replaced (medium, found twice): registerGraphics only
  reassigns the name map, so a second top-level volren/volslice sharing an id left
  the first parented under the root, still rendered + ticked, unaddressable, leaking
  a bounds signal. Now both branches removeGraphics(id) when !parent && hasGraphics,
  AFTER the volume load succeeds — matching sg.addGraphics last-wins semantics.
  New cvcgl_ariadne_realize check: two same-id volren -> first weak_ptr expires,
  last survives, getGraphics resolves to the survivor.
- volren TF `window` was bound to the density CULL (window_enabled), silently
  deleting voxels and diverging from volslice + the DSL contract (window = TF
  domain). Dropped that binding: with has_window the loader sets auto_domain=false
  so volren bakes over the control-point extent (author the domain via control-point
  values). A true arbitrary TF window for volren needs a tf_domain field in
  cvc::volren — noted as a follow-up.
- volren blank-guard ignored `unshaded`: a non-empty TF with both shaded and
  unshaded off still renders blank. Broadened guard 1 to
  isosurfaces.empty() && (tf.empty() || !(shaded||unshaded)); guard 2 unchanged.
- TF/isosurface value-domain quantities were float, narrowing the renderer's double
  domain. SceneTFPoint::value, SceneTransferFunction::window_min/max, and
  SceneIsosurface::value are now double (colors stay float); loader casts dropped.

All green: text 11 + loader 32 + runtime 11 + bind 15 + cvcgl_ariadne_realize.

+++ inc/cvc/ariadne/scene.h
@@ -15,9 +15,10 @@ namespace cvc {
 namespace ariadne {
 
 // A transfer-function control point over the RAW value domain (mirrors
-// cvc::volren::transfer_point, shared by volren + volslice). color is r,g,b,a in [0,1].
+// cvc::volren::transfer_point, shared by volren + volslice). `value` is double to
+// match the renderer's value domain; color is r,g,b,a in [0,1].
 struct SceneTFPoint {
-  float value = 0.0f;
+  double value = 0.0;
   float color[4] = {0.0f, 0.0f, 0.0f, 0.0f};
 };
 
@@ -26,13 +27,13 @@ struct SceneTransferFunction {
   std::vector<SceneTFPoint> points;
   bool auto_domain = true;   // bake over the volume's data range (vs an explicit window)
   bool has_window = false;   // explicit value window instead of the data range
-  float window_min = 0.0f, window_max = 0.0f;
+  double window_min = 0.0, window_max = 0.0; // raw value domain — double like the renderer
   bool empty() const { return points.empty(); }
 };
 
 // One volren isosurface (mirrors cvc::volren::isosurface).
 struct SceneIsosurface {
-  float value = 0.0f;
+  double value = 0.0; // raw value domain — double like cvc::volren::isosurface::value
   float opacity = 1.0f;
   float color[3] = {1.0f, 1.0f, 1.0f};
   float shininess = 10.0f;
